+++ src/main/java/net/oneandone/sushi/fs/ResourceFromModuleException.java
@@ -1,0 +1,9 @@
+package net.oneandone.sushi.fs;
+
+import java.net.URL;
+
+public class ResourceFromModuleException extends RuntimeException {
+    public ResourceFromModuleException(URL resource) {
+        super("resource from module, not in classpath: " + resource);
+    }
+}