+++ src/main/java/net/oneandone/sushi/fs/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package net.oneandone.sushi.fs;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String resource) {
+        super("resource not found in classpath: " + resource);
+    }
+}