+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   * renamed World.locateClasspathItem() to locateClasspathEntry() to make possible Java 9 upgrade problems explicit: 
     localteClasspathEntry does not locate modules
 
-* update svnkit dependency 1.8.12 to 1.8.14
+* update svnkit dependency 1.8.12 to 1.9.3
 
 * changed Serializer.serializeChildren(node/doc) to Serializer.serializeChildren(node/doc, format); the previous
   version added *some* formatting, the new version adds formatting if the second argument is true