+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * support Java 9 and 10 
   * renamed World.locateClasspathItem() to locateClasspathEntry() to make possible Java 9 upgrade problems explicit: 
-    localteClasspathEntry does not locate modules
+    localteClasspathEntry does not locate modules. The new method throws distinguished RuntimeExceptions to report
+    if a resource is not found (ResourceNotFoundException) or if a resource is from a module (ResourceFromModuleException)
 * update svnkit dependency 1.8.12 to 1.9.3
 * changed Serializer.serializeChildren(node/doc) to Serializer.serializeChildren(node/doc, format); the previous
   version added *some* formatting, the new version adds formatting if the second argument is true